(feat) remove image from list
