+++ ssu_app/templates/msg.html.twig
@@ -1,0 +1,3 @@
+<div class="col-12">
+  <p id="msg" class="text-center"></p>
+</div>
