+++ ssu_app/templates/uploader/index.html.twig
@@ -23,9 +23,7 @@
 					<div class="col-9">
 						<div id="pasteArea" class="form-control rounded-3" value="" contenteditable="true" name="Screenshot" style="width: 100%; height: 200px; border: 1px solid #ccc;"></div>
 					</div>
-					<div class="col-12">
-						<p id="msg" class="text-center"></p>
-					</div>
+					{% include "msg.html.twig" %}
 					<div class="col-12">
 						<p><a href="{{ list_of_uploaded_image }}" class="link-success link-offset-2 link-underline-opacity-25 link-underline-opacity-100-hover">Go to list of images</a></p>
 					</div>
@@ -42,6 +40,7 @@
 
 	{% block javascripts %}
     {{ parent() }}
-		<script type="text/javascript" src="{{ asset('assets/js/screenshot.js?v=' ~ date().timestamp) }}"></script>		
+		<script type="text/javascript" src="{{ asset('assets/js/main.js?v=' ~ date().timestamp) }}"></script>
+		<script type="text/javascript" src="{{ asset('assets/js/screenshot.js?v=' ~ date().timestamp) }}"></script>
 	{% endblock %}
 {% endblock %}
