+++ ssu_app/templates/uploader/list_uploaded_img.html.twig
@@ -3,6 +3,12 @@
 {% block title %}Hello UploaderController!
 {% endblock %}
 
+{% block stylesheets %}
+{{ parent() }}
+<link rel="stylesheet" href="{{ asset('assets/css/clipboardImageStyle.css') }}" >
+<link rel="stylesheet" href="{{ asset('assets/css/formStyle.css') }}" >
+{% endblock %}
+
 {% block body %}
 
 	<main>
@@ -12,6 +18,7 @@
 						<h1 class="mb-4">List of uploaded images</h1>
             <a href="{{ pathToUploadForm }}" class="link-body-emphasis link-offset-2 link-underline-opacity-25 link-underline-opacity-75-hover">Back to upload form</a>
 					</div>
+					{% include "msg.html.twig" %}
 					<div class="col-12">
             <div class="row row-cols-4 row-cols-sm-4 row-cols-md-4 g-3">
 						{% for key,value in filesInDir %}
@@ -26,10 +33,15 @@
 								</div>
 							</div>
 						{% endfor %}
+						<input id="uploaderPath" type="hidden" name="uploaderPath" value="{{ path('uploader') }}">
 					</div>
           </div>
 				</div>
 		</div>
 	</main>
-
+{% block javascripts %}
+    {{ parent() }}
+		<script type="text/javascript" src="{{ asset('assets/js/main.js?v=' ~ date().timestamp) }}"></script>
+		<script type="text/javascript" src="{{ asset('assets/js/screenshot-list.js?v=' ~ date().timestamp) }}"></script>
+	{% endblock %}
 {% endblock %}
